# mobile/agent-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Delphi client — API gateway surfaces",
    "version": "1.0.0",
    "description": "Machine-oriented OpenAPI for **AI agents only**. Not exposed to end-user browsers without operator credentials.\n\nServed from **`GET /api/openapi.json`** or alias **`GET /api/agent/openapi.json`** when the host sets **`WEB_AGENT_OPENAPI_SECRET`**; send the same value as **`X-Agent-OpenAPI-Secret`** or **`Authorization: Bearer`**. If the secret is not configured on the host, these URLs respond **404** (no document for casual visitors).\n\n**Envelopes**\n- Most catalog / mall / bet JSON: `{ data, errorCode, message, _req_id }` (numeric `errorCode`, `0` = success).\n- User-agg auth: `{ errorCode, message?, data? }` with `errorCode === 0` on success.\n\n**Auth**\n- Mall-agg (products, mall orders, checkout, points): `Authorization: Bearer <access_token>`.\n- Bet orders / place / points: `X-User-Access-Token: <access_token>` (raw JWT).\n- Bet catalog GETs (games, markets, selections): typically no user header."
  },
  "servers": [
    {
      "url": "/",
      "description": "Resolved gateway origin from runtime config (Fusio public config `data.value.host`)."
    }
  ],
...
  "paths": {
    "/api/openapi.json": {
      "get": {
        "tags": ["Meta"],
        "summary": "OpenAPI document (AI agents — credential required on host)",
        "description": "Same payload as `/api/agent/openapi.json`. No public anonymous access when `WEB_AGENT_OPENAPI_SECRET` is set; without it hosts should return 404.",
        "parameters": [
          {
            "name": "X-Agent-OpenAPI-Secret",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Must match host `WEB_AGENT_OPENAPI_SECRET` (or Bearer with same value)."
          }
        ],
        "responses": {
          "200": { "description": "Full OpenAPI 3.0 JSON" },
          "403": { "description": "Wrong or missing agent credential" },
          "404": { "description": "Host disabled OpenAPI or file missing" }
        }
      }
    },
    "/api/agent/openapi.json": {
      "get": {
        "tags": ["Meta"],
        "summary": "Alias of GET /api/openapi.json",
        "description": "Same contract and credential as `/api/openapi.json`.",
        "parameters": [
          {
            "name": "X-Agent-OpenAPI-Secret",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Must match host `WEB_AGENT_OPENAPI_SECRET` (or Bearer with same value)."
          }
        ],
        "responses": {
          "200": { "description": "Full OpenAPI 3.0 JSON" },
          "403": { "description": "Wrong or missing agent credential" },
          "404": { "description": "Host disabled OpenAPI or file missing" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "MallBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token from `POST /api/user-agg/login`"
      },
      "BetUserAccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-User-Access-Token",
        "description": "Raw JWT (`access_token` from login)."
      }
    },
...
  }
}